Ensure release process automatically publishes non-draft release (#4620)
## Changes
With the updated release automation based on the `tagging` workflow, to
complete the process of automating the full release lifecycle, the
`release` workflow should use the latest changelog entry and publish a
non-draft release. This is a small change to read the release notes from
the most recent changelog entry and supply that to goreleaser.

## Tests

- [x] Added a small unit test for the extract_release_notes.py utility
that is run as part of unit tests for the repo.

I can't easily manually test the workflow changes, so we will need to do
a small release to verify. I'll wait until a single change is merged to
test this out.

NO_CHANGELOG=true