Remove GKE test cases from databricks_mws_workspaces and databricks_mws_networks (#4625)
## Changes
GKE-based GCP workspaces are no longer supported, and all newly created
workspaces are backed by GCE. The current GKE-based network and
workspace tests can race because it is forbidden to use the same IP
ranges in two different networks. As that feature is deprecated, we
should be safe to simply remove these tests entirely.

Resolves #4618

## Tests
N/A.

NO_CHANGELOG=true

Co-authored-by: Alex Ott <alexey.ott@databricks.com>