Increase HTTP timeout to 65 seconds (#4628)
## Changes
The SDK's default HTTP timeout is 30 seconds for making a request (after
the last byte has been sent) and 30 seconds for reading the response.
This does not align with the REST API, where the HTTP timeout for many
services is closer to 60 seconds. This causes some long-running requests
to be canceled on the client side before the service finishes handling
the request. Additionally, if the server-side timeout is exceeded, an
error message is returned, but this message is currently ignored by the
SDK because it is only returned after the client-side timeout is
reached.

This PR addresses this by setting the default client-side timeout to 65
seconds for both the SDKv2 and Plugin Framework provider
implementations. If this setting is already configured, we'll respect
the configured value.

## Tests
- [x] Unit test verifying that the default timeout is set to 65 seconds
when unspecified but can be overridden to a different value.