Fix release.yml treatment of release notes (#4632)
## Changes
There is a bug in the current release job: writing out the release notes
to a file causes the following error message to be printed:
```
  ⨯ release failed after 0s                 
    error=
    │ git is in a dirty state
    │ Please check in your pipeline what can be changing the following files:
    │ ?? release_notes.md
```
Instead of writing the notes to a file, we can pipe them directly to the
`goreleaser` command as described in
https://goreleaser.com/customization/release/#custom-release-notes.

## Tests
Ran the command locally. It's hard to test this effectively. We can try
it again once this is merged.

NO_CHANGELOG=true