Use the Go SDK for `databricks_mws_workspaces` (#4633)
## Changes
Currently, many resources in the `mws` package are implemented manually
in the Terraform provider. Refactoring these to use the Go SDK reduces
the amount of code that needs to be maintained in the TF provider and
also ensures that new changes are released automatically. This PR should
be nearly identical in terms of behavior, but there are some small
variations because of subtle limitations in the Go SDK. For example, we
may need to fetch the workspace an additional time using the Go SDK
because `WaitGetWorkspaceRunning` does not return the fetched workspace
in a terminal failure state.

### Refactoring and Improvements:
* Refactored `databricks_mws_workspaces` resource and datasource, as
well as `databricks_mws_ncc_binding`, to use the Go SDK.

### Caching Mechanisms:
* Introduced caching for workspace clients in `DatabricksClient` to
improve performance and reduce redundant API calls. Added
`cachedWorkspaceClient`, `cachedWorkspaceClients`, and
`cachedAccountClient` fields along with associated methods.
[[1]](diffhunk://#diff-e24a91e4fac6650a021a72e8527d08f88b5775ee989f109f5c6ead0c3737c7a6R57-R68)
[[2]](diffhunk://#diff-e24a91e4fac6650a021a72e8527d08f88b5775ee989f109f5c6ead0c3737c7a6R104-R140)
[[3]](diffhunk://#diff-e24a91e4fac6650a021a72e8527d08f88b5775ee989f109f5c6ead0c3737c7a6R175-R180)

### Test Case Updates:
* Updated `TestDataSourceMwsWorkspaces` to use the new Go SDK mocks for
testing, replacing HTTP fixtures with mock client responses.
[[1]](diffhunk://#diff-72f2e86d45da988ff0f665f631b5bc5cc3fc42c3f27baf0074a81eb196fcbe01R4-R25)
[[2]](diffhunk://#diff-72f2e86d45da988ff0f665f631b5bc5cc3fc42c3f27baf0074a81eb196fcbe01L41-L44)
[[3]](diffhunk://#diff-72f2e86d45da988ff0f665f631b5bc5cc3fc42c3f27baf0074a81eb196fcbe01L54-R56)
* Refactored `TestMwsAccWorkspaces_TokenUpdate` to include additional
checks and new token update scenarios.
* Modified `TestResourceNccBindingCreate` and
`TestResourceNccBindingUpdate` to use mock clients for workspace
updates.
[[1]](diffhunk://#diff-f29f55efc35edd4651c26d2c0d5ce293077c0eb51fc8076e547708dee11f0c2eR5-R31)
[[2]](diffhunk://#diff-f29f55efc35edd4651c26d2c0d5ce293077c0eb51fc8076e547708dee11f0c2eL42-R49)

### Test Framework Enhancements:
* Added support for multiple mock workspace clients in the
`ResourceFixture` struct and validation logic.
[[1]](diffhunk://#diff-6b1eb785cebf167edccebd966bbd0625751fc85cea532ce6c5a3eddbca793cc6R83-R84)
[[2]](diffhunk://#diff-6b1eb785cebf167edccebd966bbd0625751fc85cea532ce6c5a3eddbca793cc6L206-R211)
[[3]](diffhunk://#diff-6b1eb785cebf167edccebd966bbd0625751fc85cea532ce6c5a3eddbca793cc6R235-R242)

## Tests
Tests run locally. Verifying with integration tests.

## TODO
- [x] Onboard `client.go` changes to code generation.
- [ ] Add back unit tests for token changes