Set release notes in a separate step during the release workflow (#4642)
## Changes

https://github.com/databricks/terraform-provider-databricks/actions/runs/14443278474
shows that the release workflow is currently failing after my changes to
publish the correct release notes. This seems to be due to an issue with
the process substitution recommended at
https://goreleaser.com/customization/release/#gitea, maybe to do with
github actions behavior. Here, I've modified the process to configure
the goreleaser arguments in a separate step.

## Tests
Ran the if/else parts on my own machine with the `github.ref`
interpolated.

NO_CHANGELOG=true