[Exporter] Correctly handle account-level identities when generating the code
When Unity Catalog metastore is attached to a workspace, the Identity Federation is
enabled on it.  With Identity Federation users, service principals and groups are coming
from the account level via assignment to a workspace.  But there is still an ability to
create workspace-level groups via API and `databricks_group` resource uses it and always
create workspace-level.  As result, we shouldn't generate resources for account-level
groups, because they will be turned into the workspace-level groups.  And due to the
limitations of APIs we can't use `databricks_permission_assignment` on workspace-level to
emulate the assignment.

See further explanation in the doc