[Exporter] Correctly handle account-level identities when generating the code (#4650)
## Changes
<!-- Summary of your changes that are easy to understand -->

When Unity Catalog metastore is attached to a workspace, the Identity
Federation is enabled on it. With Identity Federation users, service
principals, and groups are coming from the account level via assignment
to a workspace. But there is still an ability to create workspace-level
groups via API, and the `databricks_group` resource uses it and always
creates workspace-level. As a result, we shouldn't generate resources
for account-level groups because they will be turned into
workspace-level groups. Due to the limitations of APIs, we can't use
`databricks_permission_assignment` on the workspace level to emulate the
assignment.

See further explanation in the doc

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [x] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework