[Exporter] Add `Ignore` implementation for `databricks_grants` to fix issue with wrongly generated dependencies (#4661)
## Changes
<!-- Summary of your changes that are easy to understand -->

The lack of `Ignore` led to the situation when we added a `depends_on`
on grant and grants list was empty, but the reference wasn't removed
from `depends_on` list because we didn't have `Ignore` implementation.

Resolves #4025

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework