Support serverless workspaces in `databricks_mws_workspaces` (#4670)
## Changes
A new field `compute_mode` has been added to
`databricks_mws_workspaces`. This can be set to `SERVERLESS` to indicate
that a workspace should be serverless. Serverless workspaces are
documented at
https://docs.databricks.com/aws/en/admin/workspace/serverless-workspaces.
These workspaces need neither `credentials_id` or
`storage_configuration_id`, so the setup should be much easier.

## Tests
Added an integration test for creating serverless workspaces in AWS.

## Todos
- [x] Need to verify if serverless workspaces work in GCP. EDIT: not yet
supported.
- [x] Need to verify which regions users can specify for serverless
workspaces. EDIT: this will be documented shortly. I'll include a link.
- [x] The test does not work right now, I think our E2 account still
needs to be onboarded. EDIT: our account was onboarded and we verified
that this works.