[Fix] Make `spark_version` field optional to work with defaults in policies (#4643)
## Changes

A cluster policy can enforce a specific `spark_version` field and set it
as the default. This mechanism allows for a centralized choice of the
Databricks Runtime version across all jobs. To allow a job to inherit
this field from a policy, it must be configured as optional in the
schema.

The job resource referred to `JobSettings` and `JobSettingsResource`
with a mix of `js` and `jsr` variable names. This PR updates references
to `JobSettingsResource` to be called `jsr`.

## Tests

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder

---------

Co-authored-by: Tanmay Rustagi <88379306+tanmay-db@users.noreply.github.com>