[Fix] Don't use single-node cluster for `databricks_sql_permissions` (#4141)
## Changes
<!-- Summary of your changes that are easy to understand -->

Due to the internal changes in the backend API, it doesn't allow to
create non-UC single-node clusters. Until the fix is implemented on the
backend, create a cluster for setting permissions consisting of 1
worker.

Also, went through the documentation's grammar.

Resolves #4140

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [x] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK