Update minimum supported Terraform version in `README.md` (#4683)
## Changes
Currently, the minimum supported version of `terraform` with the
Databricks TF Provider is documented as v1.0. However, there is an issue
with too-strict validation of resources using SDKv2. This updates the
guidance for customers to use `terraform` from v1.1.5, which mitigates
this issue. Generally, new resources will be implemented using the
Plugin Framework, which does have strict enforcement of these bugs. Over
time, we will migrate resources to use the plugin framework, and when we
do so, we will need to address them, but for now, the status quo is that
they are tolerated. By using v1.1.5 and later, this will not result in
errors in plan/apply steps.

## Tests
No tests needed, this is a docs-only change.