[Internal] Relax the check for "parent folder doesn't exist" error (#4690)
## Changes
<!-- Summary of your changes that are easy to understand -->

Check that parent folder doesn't exist in `databricks_notebook` and
`databricks_workspace_file` is too specific and leads to an error in
staging. This PR relax this check a bit to avoid problems.

NO_CHANGELOG=true

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework