Fix use of `force` option for service principals in `databricks_git_credential` (#4704)
## Changes
<!-- Summary of your changes that are easy to understand -->

When we're trying to set git authentication for service principal, the
error message is slightly different from that for users, so I tweaked it
to work with both. Also adjusted the docs `git_username` and
`personal_access_token` aren't required for all Git providers (i.e., for
Entra ID Auth).

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [x] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework