Fix reading of `external_id` for `databricks_service_principal` (#4712)
## Changes
<!-- Summary of your changes that are easy to understand -->

During work on exporter I found that `external_id` wasn't filled for
service principals and this lead to a problem with generation of correct
code. The `databricks_user` had the correct code, so I unified the
things

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework