[Fix] Throw an error when creating `databricks_query` and `databricks_alert` with already existing names (#4697)
## Changes
<!-- Summary of your changes that are easy to understand -->

Right now, when an object with the same name already exists in a
workspace, the Query and Alert APIs just create an object with a
different name (by appending `(1)` or something like that). This leads
to permanent configuration drift as TF will try to rename the object to
the name specified in the HCL.

P.S. Waiting for support this flag in the Update API as well...

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework

---------

Co-authored-by: vuong-nguyen <44292934+nkvuong@users.noreply.github.com>