Add `schema_id` to the output of `databricks_schema` resource (#4773)
## Changes
Schema now exports a stable UUID identifier in the `schema_id` field.
This PR adds this to the model for schema as a computed attribute.

## Tests
Manually verified that the schema_id was populated in state in GCP. I
don't want to add a test for this per se since it's possible that it
isn't populated in all environments (yet).

---------

Co-authored-by: Alex Ott <alexey.ott@databricks.com>