[DOCS] added example of granting Account Admin role to a service principal (#4807)
## Changes
<!-- Summary of your changes that are easy to understand -->
This PR makes changes to docs only, no code changes. The
`databricks_service_principal_role` is capable of giving account admin
role to a service principal; however, this was not very clear in the
docs whereas we do have an example given in `databricks_user_role` for
users.

I added an example demonstrating this capability of setting `role =
"account_admin"` and a note that it can only be used with an
account-level provider.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [X] relevant change in `docs/` folder