Use single-node cluster for `databricks_sql_permissions`
In 27ad78d, we modified the default cluster launched by the `databricks_sql_permissions` resource to use a cluster with a worker node. This was done to mitigate an issue where it was no longer possible to create single-node clusters with legacy Table ACLs. Since the start of May, single-node `USER_ISOLATION` clusters are supported and can be used to manage Table ACLs. This PR changes this resource to use such a cluster going forward.