Use single-node cluster for `databricks_sql_permissions` (#4813)
## Changes
<!-- Summary of your changes that are easy to understand -->

In 27ad78d, we modified the default cluster launched by the
`databricks_sql_permissions` resource to use a cluster with a worker
node. This was done to mitigate an issue where it was no longer possible
to create single-node clusters with legacy Table ACLs. Since the start
of May, single-node `USER_ISOLATION` clusters are supported and can be
used to manage Table ACLs. This PR changes this resource to use such a
cluster going forward.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [x] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework