Add documentation for single-node clusters in `databricks_cluster` (#4817)
## Changes
The `is_single_node` field was recently introduced for clusters to
simplify creation of single-node clusters. This PR updates the Terraform
documentation for `databricks_cluster` to show how to configure the
cluster resource using this new flag. This simplifies the process from
before, which required a number of settings to be configured
simultaneously.

## Tests
Manually tested by creating a cluster with the documented template.