Add synthetic field to control `apply_policy_default_values` (#4834)
## Changes

This commit introduces a new synthetic field,
`__apply_policy_default_values_allow_list`, to the job resource schema.
This field is a list of strings representing the cluster config
attributes that the user explicitly wants to manage. During create and
update operations, only the fields listed in this attribute will be
included in the API payload for the job's cluster configuration. All
other fields will be omitted, allowing the backend to apply or update
cluster policy defaults as needed (when `apply_policy_default_values =
true`).

This approach works around a limitation in the legacy Terraform Plugin
SDK, where it is not possible to distinguish between user-specified and
server-defaulted fields after the initial create. By requiring users to
explicitly list the fields they wish to control, we enable them to reset
cluster config fields to new policy defaults by removing them from both
the config and the synthetic list.

- Adds `__apply_policy_default_values_allow_list` to the job resource
schema.
- Filters cluster config fields in the API payload based on this list.
- Documents the new behavior and usage pattern for users.

## Tests

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [x] has entry in `NEXT_CHANGELOG.md` file

---------

Co-authored-by: Omer Lachish <289488+rauchy@users.noreply.github.com>