[Doc] Update disable_legacy_features setting documentation  (#4884)
## Changes
<!-- Summary of your changes that are easy to understand -->
This PR updates the documentation of the
`disable_legacy_features_setting` resource to remove the mention of it
being in Private Preview, as it will go Public preview next week. It
also renames the `disable_default_legacy_access.md` file to
`disable_legacy_access.md`, to reflect the actual name of the resource.
## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [x] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [x] has entry in `NEXT_CHANGELOG.md` file

Co-authored-by: Alex Ott <alexey.ott@databricks.com>