Do not set `run_as` from `run_as_user_name` in DLT pipelines (#4886)
## Changes

This PR removes custom handling that we have for run_as for DLT:
1. Remove the computed annotation for run_as in Terraform. It's not a
server owned field and is instead meant to be set by the client.
2. Do not read server-side `run_as_user_name` values and set it in
run_as.

This PR reverts some code that was originally introduced in:
1. https://github.com/databricks/terraform-provider-databricks/pull/4456
2. https://github.com/databricks/terraform-provider-databricks/pull/4470

This custom code was added for `run_as` because originally the API never
returned values for the run_as field in the GET API call, so Terraform
had to depend on values from `run_as_user_name`. Since the backend was
patched to return values for run_as, we no longer need the custom code
to read `run_as_user_name`.

Removing this custom code fixes the caching of run_as in Terraform
state. Before, removing run_as from the TF config would have been a
noop. Now, TF will correctly return an error when a user tries to do so.

NO_CHANGELOG=true (The feature is in PrPr)

Resolves #4897

## Tests
Manually created a pipeline with an old version of the provider and
checked that upgrading to this PR does not cause Terraform apply to
fail. Both when run_as is set and when run is not set.

Existing acceptance tests: TestAccPipelineRunAsUser,
TestUcAccPipelineRunAsServicePrincipal, TestUcAccPipelineRunAsMutations

Updated TestUcAccPipelineRunAsMutations to include two new cases:
1. Setting run_as to a pipeline that did not have one configured before
works.
2. Removing run_as causes an error.

-------------------------

Note: This is a breaking change. If a user had a run_as configured on
their DLT pipeline, but then removed it at some point they'll start to
see the following error when they update their Terraform version:
```
databricks_pipeline.basic_pipeline: Modifying... [id=6f832064-c8b8-4a78-8f6f-f8da5014913c]
╷
│ Error: cannot update pipeline: Once configured, run_as of the pipeline can not be set to null. Please set run_as to fa95fb60-f59b-4b78-832d-83f41da6f284 to preserve the pipeline's run_as
│ 
│   with databricks_pipeline.basic_pipeline,
│   on main.tf line 13, in resource "databricks_pipeline" "basic_pipeline":
│   13: resource "databricks_pipeline" "basic_pipeline" {
│ 
```

To fix they can simply set run_as back again on their pipelines. This is
fine because run_as is still in private preview.