[Feature] Enable the option to add `comment` when defining a new `databricks_share` (#4802)
## Changes
Added the option for the user to define a `comment` when creating
`databricks_share` which will be displayed in Databricks UI for defined
share.

Additionally, made `share.md` documentation more consistent with
Terraform style.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [x] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework

---------

Co-authored-by: Alex Ott <alexey.ott@databricks.com>
Co-authored-by: Alex Ott <alexott@gmail.com>