Fix schema mismatch bug in `databricks_functions` data source (#4902)
## Changes
<!-- Summary of your changes that are easy to understand -->

The cause of the bug was that SDKv2-compatible struct was used with the
plugin framework. Switching to use correct struct fixed the problem.

Bug wasn't found previously because we don't have integration test for
it as we don't have `databricks_function` resource yet.

Resolves #4365


## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [x] using Go SDK
- [x] using TF Plugin Framework
- [x] has entry in `NEXT_CHANGELOG.md` file