DOCS: Clarify grant vs grants (#4935)
## Changes
It was a bit vague to understand the difference between grant and grants
based on the current description. The text made sense after you
understood the difference, but this small change will make it instantly
understandable.

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [x] has entry in `NEXT_CHANGELOG.md` file