DOCS: Fix broken link in index (#4949)
As seen in the [current version
1.87.1](https://registry.terraform.io/providers/databricks/databricks/1.87.1/docs),
this link is broken. Works in GitHub, but not on on the Terraform
Registry.

## Changes
<!-- Summary of your changes that are easy to understand -->

In docs/index.md

it was (../docs/guides/aws-workspace.md), but I changed it to
(guides/aws-workspace.md). This is also more in line with the other
links that do work.

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [ ] has entry in `NEXT_CHANGELOG.md` file

---------

Co-authored-by: Alex Ott <alexey.ott@databricks.com>