Fix regression with `databricks_group` data source introduced by a recent change
It looks like that some users are using `databricks_group` data source as non-admin
users (in some cases - incorrectly).  Recent change in SCIM API required to change from
List to List for `id` + Get to retrieve group members, but it had an unexpected effect
that Get for non-admin users throws an error as we request for `externalId` attribute.

I got clarification from the SCIM team that changes in the `members` retrieval is
affecting only account-level API, and not workspace-level API.  So this PR reimplements
the required changes as following:

- on account level we do `List` for group ID by name, followed by `Get` to retrieve
  `members`, `externalId`, etc.
- on workspace level we use the old inmplmentation of `List` + all attributes

Resolves #4994