Fix regression with `databricks_group` data source introduced by a recent change (#4995)
## Changes
<!-- Summary of your changes that are easy to understand -->

It looks like some users are using the `databricks_group` data source as
non-admin users (in some cases, incorrectly). Recent change in SCIM API
required to change from List to List for `id` + Get to retrieve group
members, but it had an unexpected effect that Get for non-admin users
throws an error as we request for `externalId` attribute.

I got clarification from the SCIM team that changes in the `members`
retrieval are affecting only the account-level API, and not the
workspace-level API. So this PR reimplements the required changes as
follows:

- on account level we do `List` for group ID by name, followed by `Get`
to retrieve `members`, `externalId`, etc.
- on workspace level we use the old implementation of `List` + default
attributes returned by filter operation

Resolves #4994
Resolves #4996


## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [x] has entry in `NEXT_CHANGELOG.md` file