[Internal] Add integration test to check re-ordering objects in `databricks_share_pluginframework ` (#5085)
## Changes
<!-- Summary of your changes that are easy to understand -->
Add integration test to verify that changing the order of objects in a
share in config leads to non-empty plan. This is the same behaviour that
we have in SDKv2.

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->
Integration test

NO_CHANGELOG=true