Document `git_email` in `databricks_git_credential` resource (#5099)
## Changes
<!-- Summary of your changes that are easy to understand -->

`git_username` is required when `git_email` is specified, but `git_email
isn't required with `git_username`. However Terraform SDKv2 doesn't
allow to express this kind of dependency, so I didn't modify the code
(it should be possible with the Plugin Framework).

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [x] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [x] has entry in `NEXT_CHANGELOG.md` file