Fix `ADMIN` assignment in `databricks_permission_assignment` resource (#5109)
## Changes
<!-- Summary of your changes that are easy to understand -->

It was found that assignment by name works only for the `USER` role, so
the direct assignment as `ADMIN` fails. This PR fixes this issue by
doing assignment as `USER` first, and then changing it to `ADMIN`.

Resolves #5106


## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [x] has entry in `NEXT_CHANGELOG.md` file