[Documentation] Remove `METRIC_VIEW` from `sql_table` resource (#5135)
## Changes
<!-- Summary of your changes that are easy to understand -->
Remove mention of METRIC_VIEW. This is causing confusion. 

From Daniel: 
The `databricks_sql_table` resource is able to manage the lifecycle of
tables (managed and external) and views. The docs erroneously included
metric views. We've chosen deliberately not to support metric views due
to the existing challenges in supporting tables/views across SQL DDL and
TF semantics.

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->
N/A