Ensure `workload_size` is marked as non-computed for `databricks_model_serving` (#5111)
## Changes
`workload_size` is currently marked as `optional` and `computed`. As a
result, if a user sets a model serving endpoint using `workload_size`
initially, then removes it in favor of
`min_provisioned_concurrency`/`max_provisioned_concurrency`,
`workload_size` remains in the state. As a result, it is also included
in the request sent to the REST API, which is invalid.

By marking this field as non-computed, it should be removed from the
plan when a user removes it from their config. This means it won't be
sent to the REST API.

## Tests
- [ ] Manual test: created a model serving endpoint with `workload_size`
set, then removed it and set min/max provisioned concurrency fields, and
the apply succeeded.

---------

Co-authored-by: Alex Ott <alexey.ott@databricks.com>