[Internal] Skip scim user and group role tests as on non-Aws environments (#5153)
## Changes
<!-- Summary of your changes that are easy to understand -->
These tests started fail on Saturday (passed on Friday) with no change
in resources for Azure and GCP environments. This could be an API issue,
running these tests on AWS to unblock PRs to merge since the failure is
happening on main branch.

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->
CI is green
NO_CHANGELOG=true