Document the `enabled` field of `data_classification_config` in `databricks_quality_monitor` (#5160)
## Changes
The structure `data_classification_config` in
`databricks_quality_monitor` is documented, but its fields are not.
Currently it has only one field, `enabled`, which is a boolean. This PR
adds documentation for this field. The comment is taken from the
internal API specification.

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [ ] has entry in `NEXT_CHANGELOG.md` file

NO_CHANGELOG=true