[Internal] Try to fix settings tests (#5231)
## Changes
<!-- Summary of your changes that are easy to understand -->

It looks like that we can use Get directly, without need to specify the
eTag

It also includes a workaround to the catalog resource because the
backend now returns a PO flag that fails on the update.

NO_CHANGELOG=true


## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [ ] has entry in `NEXT_CHANGELOG.md` file