[Documentation] Update documentation for data object type for share resource (#5244)
## Changes
<!-- Summary of your changes that are easy to understand -->
The documentation for data object type is out of date as it's manually
maintained. Updating it to include types that are supported eg: `VOLUME`
. Also added `content` and `partition` as part of documentation.

Note: List is based on the changes merged internally.

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->
N/A