[Fix] Reset PO flag for non-managed UC Catalogs (#5260)
## Changes
<!-- Summary of your changes that are easy to understand -->

In some cases, UC API returned the PO flag and then we submitted it as
part of the Update operation. But now UC API has changed the behavior
and rejects the PO flag.

Fixes #5257


## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [x] has entry in `NEXT_CHANGELOG.md` file