[Feature] Support for payload bigger than 10Mb in `databricks_workspace_file`
Until this PR, the `databricks_workspace_file` resource was using `Import` and `Export`
functions that has a 10Mb limit on uploading or downloading the data.  Instead of these
functions, we should use `Upload` and `Download` functions that support bigger payload (up
to 500Mb).

The corresponding changes were made in TF Exporter as well.